+++ includes/Header/Body_Class.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.1.0
+ * @version  1.1.3
  */
 
 namespace WebManDesign\Zooey\Header;
@@ -54,7 +54,7 @@ class Body_Class implements Component_Interface {
 	 * HTML body classes.
 	 *
 	 * @since    1.0.0
-	 * @version  1.1.0
+	 * @version  1.1.3
 	 *
 	 * @param  array $classes
 	 *
@@ -107,6 +107,15 @@ class Body_Class implements Component_Interface {
 			) {
 				$classes[] = 'search-generic';
 			}
+
+			/**
+			 * Temporary a11y fix to prevent broken focus trap on mobile menu submenus.
+			 *
+			 * @todo  Remove this once WordPress fixes Navigation block JS.
+			 * @link  https://github.com/WordPress/gutenberg/issues/63033
+			 * @see   assets/scss/blocks/_navigation-mobile.scss
+			**/
+			$classes[] = 'navigation-a11y-fix';
 
 			// Sort classes alphabetically.
 			asort( $classes );